fastSPDMome: separate real temporary from persistent files.
* Keep all files in /tmp, and only move the actual files that need to be served
into a persistent directory such as /var/lib/appname. For example, log files
are read from /tmp and not served at all. At the moment, this is only the
imstres image.
* Using /tmp as much as possible, also means that whatever happens, we can
count on the system to clean afterwards (we will still clean up /tmp after
ourselves in if all goes well).