nd-safir.py: keep connection alive during long process.
* omero/nd-safir.py: large images can take a long time (even hours) to process
after which Omero would have killed the connection. The KeepAlive method
allows to specify a time, but does not overload the server settings
'omero.sessions.timeout'. This solves this problem by creating a child process
for the processing, while the parent sleeps and only sends KeepAlive signals
to the server. This method also adds a timeout option, after which the process
is killed.