ndsafir.py: replace dv_utils static methods with member methods.
* Most of the static methods of dv_utils would accept a ImagreWrapper,
get the header of the filw, and then check for values in that header.
Rather than getting the header each time, which causes query to omero,
simply store the header in an attribute and turn them into a typical
class instead of utils class.
* Also create a @if_maybe_mrc to skip the check if the file was not
even long enough to be a MRC file.