lothar-nna: make use of bioformats for input check.
We use bioformats to check if the image really is a Z stack with 3 channels.
This saves the current issue of accidentally opening a bad image and trashing
the whole thing.

Idea is to also use it to get pixel resolution (at the moment are hardcoded),
and automatically pick the DAPI channel.