nna: remove requirement on imagej package and qt graphics toolkit.
We do not really make use of the imagej package (which doesn't even
exist publicly).  We only need it as an easy way to check that we
have ij.jar on the path.  Replace that.

Also, we need qt over fltk because the later does not support print
of invisible plots.  However, many users do not have qt available.
So gnuplot should work as well.